Revert "Gilded Rose stories tests" as it was done out of course.
Stories are in separate test file.
